refactors and organizes overlay logic. implements extensible render selector
