+++ PhotoPoints/Controller/ImageManager.swift
@@ -74,5 +74,5 @@ class ImageManager {
             }
         }
     }
-    
+
 }
