+++ PhotoPoints/View/Map/OverlayManager.swift
@@ -8,66 +8,76 @@
 
 import MapKit
 
+// overlay extensions to apply distinct formatting for each feature type
+class BoundaryGon: MKPolygon {}
+class WetlandGon: MKPolygon {}
+class StreamLine: MKPolyline {}
+class TrailLine: MKPolyline {}
+
+// renderer extensions applying distinct formatting
+class BoundaryRenderer: MKPolygonRenderer {
+    override init(overlay: MKOverlay) {
+        super.init(overlay: overlay)
+        strokeColor = .gray
+        lineWidth = 3
+    }
+}
+
+class WetlandRenderer: MKPolygonRenderer {
+    override init(overlay: MKOverlay) {
+        super.init(overlay: overlay)
+        strokeColor = .systemBlue
+        fillColor = .blue
+        lineWidth = 3
+    }
+}
+
+class TrailRenderer: MKPolylineRenderer {
+    override init(overlay: MKOverlay) {
+        super.init(overlay: overlay)
+        strokeColor = .brown
+        lineWidth = 5
+        lineDashPattern = [6]
+    }
+}
+
+class StreamRenderer: MKPolylineRenderer {
+    override init(overlay: MKOverlay) {
+        super.init(overlay: overlay)
+        strokeColor = .systemBlue
+        lineWidth = 3
+    }
+}
+
 class OverlayManager {
-    
-    lazy var renderers = [
-        NSStringFromClass(BoundaryGon.self) : boundaryRenderer,
-        NSStringFromClass(WetlandGon.self) : wetlandRenderer,
-        NSStringFromClass(StreamLine.self) : streamRenderer,
-        NSStringFromClass(TrailLine.self) : trailRenderer,
+
+    // dictionary to look up appropriate renderer for given overlay type
+    var renderers = [
+        NSStringFromClass(BoundaryGon.self) : BoundaryRenderer.self,
+        NSStringFromClass(WetlandGon.self) : WetlandRenderer.self,
+        NSStringFromClass(StreamLine.self) : StreamRenderer.self,
+        NSStringFromClass(TrailLine.self) : TrailRenderer.self,
     ]
     
-    let boundaryRenderer: MKPolygonRenderer = {
-        let renderer = MKPolygonRenderer()
-        renderer.strokeColor = .gray
-        renderer.lineWidth = 3
-        return renderer
-    }()
-    
-    let wetlandRenderer: MKPolygonRenderer = {
-        let renderer = MKPolygonRenderer()
-        renderer.strokeColor = .systemBlue
-        renderer.fillColor = .blue
-        renderer.lineWidth = 3
-        return renderer
-    }()
-    
-    let trailRenderer: MKPolylineRenderer = {
-        let renderer = MKPolylineRenderer()
-        renderer.strokeColor = .brown
-        renderer.lineWidth = 5
-        renderer.lineDashPattern = [6]
-        return renderer
-    }()
-    
-    let streamRenderer: MKPolylineRenderer = {
-        let renderer = MKPolylineRenderer()
-        renderer.strokeColor = .systemBlue
-        renderer.lineWidth = 3
-        return renderer
-    }()
-    
-    //Adds map annotations to aide the user when navigating the forest
+    // adds different types of overlays to map
     func addOverlays(to mapView: MKMapView) {
         
-        // forest boundary
-        let boundaryGon = BoundaryGon(coordinates: Boundary.boundary, count: Boundary.boundary.count)
-        mapView.addOverlay(boundaryGon)
+        for boundary in Overlays.boundary {
+            let boundaryGon = BoundaryGon(coordinates: boundary, count: boundary.count)
+            mapView.addOverlay(boundaryGon)
+        }
         
-        // streams
-        for stream in Streams.streams {
+        for stream in Overlays.streams {
             let streamLine = StreamLine(coordinates: stream, count: stream.count)
             mapView.addOverlay(streamLine)
         }
         
-        // wetlands
-        for wetland in Wetlands.wetlands {
+        for wetland in Overlays.wetlands {
             let wetLandGon = WetlandGon(coordinates: wetland, count: wetland.count)
             mapView.addOverlay(wetLandGon)
         }
         
-        // trails
-        for trail in Trails.trails {
+        for trail in Overlays.trails {
             let trailLine = TrailLine(coordinates: trail, count: trail.count)
             mapView.addOverlay(trailLine)
         }
