--- PhotoPoints/View/Map/Overlays/Boundary.swift
@@ -1,29 +1,0 @@
-//
-//  Boundary.swift
-//  PhotoPoints
-//
-//  Created by Clay Suttner on 12/22/20.
-//  Copyright © 2020 Cascadia College. All rights reserved.
-//
-
-import Foundation
-import MapKit
-
-class BoundaryGon: MKPolygon {}
-
-class Boundary {
-    
-    static let boundary = [
-        CLLocationCoordinate2D(latitude: 47.774522, longitude: -122.196209),
-        CLLocationCoordinate2D(latitude: 47.774514, longitude: -122.191098),
-        CLLocationCoordinate2D(latitude: 47.774979, longitude: -122.191186),
-        CLLocationCoordinate2D(latitude: 47.775426, longitude: -122.191302),
-        CLLocationCoordinate2D(latitude: 47.776040, longitude: -122.191517),
-        CLLocationCoordinate2D(latitude: 47.776081, longitude: -122.191358),
-        CLLocationCoordinate2D(latitude: 47.776249, longitude: -122.191419),
-        CLLocationCoordinate2D(latitude: 47.776814, longitude: -122.191702),
-        CLLocationCoordinate2D(latitude: 47.783690, longitude: -122.196068),
-        CLLocationCoordinate2D(latitude: 47.783536, longitude: -122.196771),
-        CLLocationCoordinate2D(latitude: 47.777843, longitude: -122.196329),
-    ]
-}
